Fixed issues with recent merges.
Fixed files that were not properly updated because of more recent changes than when the decimal precision edits were requested to merge. Also fixed a duplicate fill preference section in the Settings view. Also added test suites made in the project to the Vitest config so they will be run with a full-project test run.